src/requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django==4.2.16
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
geographiclib==2.0
geopy==2.4.1
psycopg==3.2.9
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.1
